[GraphTrainer] Move codegen tests to dedicated test_codegen.py
Extract TestCustomCodegenPass from test_passes.py into its own file
while keeping it discoverable via test_passes.py through re-import.
